chore(translations): move files to match v3 build
