+++ packages/translations/src/Provider/index.js
@@ -1,0 +1,1 @@
+export { default } from './Provider';
