+++ packages/translations/src/intlHelper/index.js
@@ -1,0 +1,1 @@
+export { default } from './intlHelper';
