+++ packages/translations/src/intlHelper/intlHelper.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import IntlProvider from './Provider';
+import IntlProvider from '../Provider';
 
 const intlHelper = (message, settings) => <IntlProvider { ...settings }>{message}</IntlProvider>;
