--- packages/translations/src/translatedMessages.js
@@ -1,1 +1,0 @@
-export { default as translatedMessages } from '../../../locales/data.json';
