+++ packages/translations/src/translatedMessages/translatedMessages.js
@@ -1,0 +1,1 @@
+export { default as translatedMessages } from '../locales/data.json';
